requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
# created-by: conda 24.9.2
absl-py=2.1.0=pypi_0
accelerate=1.3.0=pypi_0
aiohappyeyeballs=2.4.6=pypi_0
aiohttp=3.11.12=pypi_0
aiosignal=1.3.2=pypi_0
albucore=0.0.23=pypi_0
albumentations=1.4.24=pypi_0
aliyun-mns-sdk=1.2.2=pypi_0
aliyun-python-sdk-core=2.16.0=pypi_0
aliyun-python-sdk-core-v3=2.13.33=pypi_0
aliyun-python-sdk-kms=2.16.5=pypi_0
annotated-types=0.7.0=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
anyio=4.8.0=pypi_0
astor=0.8.1=pypi_0
async-timeout=5.0.1=pypi_0
attrdict=2.0.1=pypi_0
attrs=25.1.0=pypi_0
babel=2.17.0=pypi_0
bce-python-sdk=0.9.29=pypi_0
beautifulsoup4=4.13.3=pypi_0
black=25.1.0=pypi_0
blinker=1.9.0=pypi_0
boto3=1.36.21=pypi_0
botocore=1.36.21=pypi_0
braceexpand=0.1.7=pypi_0
brotli=1.1.0=pypi_0
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.12.31=hecd8cb5_0
cachetools=5.5.1=pypi_0
certifi=2025.1.31=pypi_0
cffi=1.17.1=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
cloudpickle=3.1.1=pypi_0
coloredlogs=15.0.1=pypi_0
colorlog=6.9.0=pypi_0
contourpy=1.3.1=pypi_0
crcmod=1.7=pypi_0
cryptography=44.0.1=pypi_0
cssselect=1.2.0=pypi_0
cssutils=2.11.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.12=pypi_0
datasets=3.3.0=pypi_0
decorator=5.1.1=pypi_0
detectron2=0.6=pypi_0
dill=0.3.8=pypi_0
distro=1.9.0=pypi_0
doclayout-yolo=0.0.2b1=pypi_0
einops=0.8.1=pypi_0
et-xmlfile=2.0.0=pypi_0
evaluate=0.4.3=pypi_0
exceptiongroup=1.2.2=pypi_0
fairscale=0.4.13=pypi_0
fast-langdetect=0.2.5=pypi_0
fasttext-predict=0.9.2.4=pypi_0
filelock=3.17.0=pypi_0
fire=0.7.0=pypi_0
flask=3.1.0=pypi_0
flask-babel=4.0.0=pypi_0
flatbuffers=25.2.10=pypi_0
fonttools=4.56.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2024.12.0=pypi_0
ftfy=6.3.1=pypi_0
future=1.0.0=pypi_0
fvcore=0.1.5.post20221221=pypi_0
grpcio=1.70.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.7=pypi_0
httpx=0.28.1=pypi_0
huggingface-hub=0.28.1=pypi_0
humanfriendly=10.0=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
imgaug=0.4.0=pypi_0
iniconfig=2.0.0=pypi_0
iopath=0.1.9=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.5=pypi_0
jiter=0.8.2=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
libffi=3.4.4=hecd8cb5_1
lmdb=1.6.2=pypi_0
loguru=0.7.3=pypi_0
lxml=5.3.1=pypi_0
magic-pdf=1.1.0=pypi_0
markdown=3.7=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.0=pypi_0
more-itertools=10.6.0=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
multiprocess=0.70.16=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=hcec6c5f_0
networkx=3.4.2=pypi_0
numpy=1.26.4=pypi_0
omegaconf=2.3.0=pypi_0
onnxruntime=1.20.1=pypi_0
openai=1.63.0=pypi_0
opencv-contrib-python=4.6.0.66=pypi_0
opencv-python=4.6.0.66=pypi_0
opencv-python-headless=4.11.0.86=pypi_0
openpyxl=3.1.5=pypi_0
openssl=3.0.15=h46256e1_0
opt-einsum=3.3.0=pypi_0
oss2=2.19.1=pypi_0
packaging=24.2=pypi_0
paddleocr=2.7.3=pypi_0
paddlepaddle=2.6.1=pypi_0
pandas=2.2.3=pypi_0
pathspec=0.12.1=pypi_0
pdf2docx=0.5.8=pypi_0
pdfminer-six=20231228=pypi_0
pillow=11.1.0=pypi_0
pip=25.0=py310hecd8cb5_0
platformdirs=4.3.6=pypi_0
pluggy=1.5.0=pypi_0
portalocker=3.1.1=pypi_0
premailer=3.10.0=pypi_0
propcache=0.2.1=pypi_0
protobuf=5.29.3=pypi_0
psutil=7.0.0=pypi_0
py-cpuinfo=9.0.0=pypi_0
pyarrow=19.0.0=pypi_0
pyclipper=1.3.0.post6=pypi_0
pycocotools=2.0.8=pypi_0
pycparser=2.22=pypi_0
pycryptodome=3.21.0=pypi_0
pydantic=2.10.6=pypi_0
pydantic-core=2.27.2=pypi_0
pymupdf=1.24.14=pypi_0
pyparsing=3.2.1=pypi_0
pytest=8.3.4=pypi_0
python=3.10.16=hce00570_1
python-dateutil=2.9.0.post0=pypi_0
python-docx=1.1.2=pypi_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
rapid-table=1.0.3=pypi_0
rapidfuzz=3.12.1=pypi_0
rapidocr-onnxruntime=1.4.4=pypi_0
rapidocr-paddle=1.4.5=pypi_0
rarfile=4.2=pypi_0
readline=8.2=hca72f7f_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
robust-downloader=0.0.2=pypi_0
s3transfer=0.11.2=pypi_0
safetensors=0.5.2=pypi_0
scikit-image=0.25.1=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
seaborn=0.13.2=pypi_0
setuptools=75.8.0=py310hecd8cb5_0
shapely=2.0.7=pypi_0
simsimd=6.2.1=pypi_0
six=1.17.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.6=pypi_0
sqlite=3.45.3=h6c40b1e_0
stringzilla=3.11.3=pypi_0
struct-eqtable=0.3.2=pypi_0
sympy=1.13.3=pypi_0
tabulate=0.9.0=pypi_0
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
termcolor=2.5.0=pypi_0
thop=0.1.1-2209072238=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2025.1.10=pypi_0
timm=0.9.16=pypi_0
tk=8.6.14=h4d00af3_0
tokenizers=0.19.1=pypi_0
tomli=2.2.1=pypi_0
torch=2.2.2=pypi_0
torchvision=0.17.2=pypi_0
tqdm=4.67.1=pypi_0
transformers=4.42.4=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
ultralytics=8.3.75=pypi_0
ultralytics-thop=2.0.14=pypi_0
unimernet=0.2.3=pypi_0
urllib3=2.3.0=pypi_0
visualdl=2.5.3=pypi_0
wand=0.6.13=pypi_0
wcwidth=0.2.13=pypi_0
webdataset=0.2.111=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py310hecd8cb5_0
xxhash=3.5.0=pypi_0
xz=5.6.4=h46256e1_1
yacs=0.1.8=pypi_0
yarl=1.18.3=pypi_0
zlib=1.2.13=h4b97444_1
